fix(politics-tracker): update latest election data handle
