+++ packages/politics-tracker/utils/politic.ts
@@ -1,8 +1,14 @@
+import moment from 'moment'
+
 import type {
   DraftPolitic,
   DraftPoliticForModification,
 } from '~/components/politics/politic-form'
-import type { ElectionData, ElectionDataForPerson } from '~/types/politics'
+import type {
+  ElectionData,
+  ElectionDataForPerson,
+  PersonElectionData,
+} from '~/types/politics'
 
 import { isTypeOfOneFromCouple } from './utils'
 
@@ -41,8 +47,36 @@ function isDraftPoliticForModification(
   )
 }
 
+function getLastestElectionData<T extends PersonElectionData>(
+  previous: T,
+  current: T
+): T {
+  const latest = previous.election
+  const election = current.election
+
+  if (election && latest) {
+    const latestTime = moment()
+      .year(latest.election_year_year)
+      .month(latest.election_year_month - 1)
+      .date(latest.election_year_day)
+      .unix()
+    const currentTime = moment()
+      .year(election.election_year_year)
+      .month(election.election_year_month - 1)
+      .date(election.election_year_day)
+      .unix()
+    if (currentTime > latestTime) {
+      return current
+    }
+  } else if (election) {
+    return current
+  }
+  return previous
+}
+
 export {
   checkIsPartyPage,
+  getLastestElectionData,
   isDraftPoliticForModification,
   isElectionDataForPerson,
 }
